+++ doc/releases/changelog-dev.md
@@ -949,6 +949,9 @@
   parts of the algorithm.
   [(#7681)](https://github.com/PennyLaneAI/pennylane/pull/7681)
 
+* Optimization for :func`pertrubation_error` which groups commutators by linearity.
+  [(#7790)][https://github.com/PennyLaneAI/pennylane/pull/7790]
+
 * Fixed missing table descriptions for :class:`qml.FromBloq <pennylane.FromBloq>`,
   :func:`qml.qchem.two_particle <pennylane.qchem.two_particle>`,
   and :class:`qml.ParticleConservingU2 <pennylane.ParticleConservingU2>`.
