+++ pennylane/labs/tests/trotter_error/product_formulas/test_error.py
@@ -22,6 +22,7 @@ from pennylane.labs.trotter_error import (
     perturbation_error,
     vibrational_fragments,
 )
+from pennylane.labs.trotter_error.product_formulas.error import _group_sums
 
 
 @pytest.mark.parametrize(
@@ -97,3 +98,21 @@ def test_perturbation_error_invalid_parallel_mode():
             backend="mp_pool",
             parallel_mode="invalid_mode",
         )
+
+
+@pytest.mark.parametrize(
+    "term_dict, expected",
+    [
+        (
+            [{("A",): 5}, {("X", "A", "B"): 4, ("Y", "A", "B"): 3}],
+            [(frozenset({("X", 4), ("Y", 3)}), "A", "B")],
+        ),
+        (
+            [{("A",): 5}, {("X", "A", "B"): 4, ("Y", "A", "C"): 3}],
+            [(frozenset({("X", 4)}), "A", "B"), (frozenset({("Y", 3)}), "A", "C")],
+        ),
+    ],
+)
+def test_group_sums(term_dict, expected):
+    """Test the private _group_sums method"""
+    assert _group_sums(term_dict) == expected
